[rocjitsu] Move kernel_symbol to code/ layer
Move find_kernel_symbol from vm/amdgpu/ to code/ as requested in PR
review — it's generic ELF parsing with no vm/amdgpu dependencies.

- Add PT_DYNAMIC, PT_NOTE, PT_LOAD, DT_* constants and Elf64_Dyn to
  amdgpu_elf.h so kernel_symbol.cpp uses project types instead of <elf.h>
- Add missing rj_code.h include to amdgpu_elf.h (elf_mach_for_arch uses
  rj_code_arch_t)
- Update namespace from rocjitsu::amdgpu to rocjitsu

Tested: full build clean, all 370 tests pass.
